Fibonacci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -12,9 +12,9 @@ int main(void)
 int i;
 int num1 = 1;
 int num2 = 2;
-long long int nextnum = num1 + num2;
+int nextnum = num1 + num2;
 
-printf("%i, %i, %lli", num1, num2, nextnum);
+printf("%i, %i, %li", num1, num2, nextnum);
 
 for (i = 4; i <= 50; i++)
 {
@@ -22,7 +22,7 @@ num1 = num2;
 num2 = nextnum;
 nextnum = num1 + num2;
 
-printf(", %lli", nextnum);
+printf(", %i", nextnum);
 }
 
 printf("\n");
